Rename to date-format.ts
